// components/pages/Order/OrderNowForm/interfaces.ts
export interface IOrderFormItem {
  name: string;
  labelText: string;
  placeholder: string;
  isTextArea?: boolean;
}

export interface ICustomizeForm {
  minDR: number;
  linkNumber: number;
  minTraffic: number;
  backLinkAudit: boolean;
  keywordAnalysis: boolean;
  optimizeArticle: boolean;
  seoAuditAndConsultation: boolean;
  linkBuildingAndStrategyPlanning: boolean;
}